Just a little code formatting and redesigned. added lombok and started upgrading to Google Guava Library.
