+++ src/main/java/com/xxmicloxx/NoteBlockAPI/Note.java
@@ -1,28 +1,14 @@
 package com.xxmicloxx.NoteBlockAPI;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
 public class Note {
-
-    private byte instrument;
-    private byte key;
-
-    public Note(byte instrument, byte key) {
-        this.instrument = instrument;
-        this.key = key;
-    }
-
-    public byte getInstrument() {
-        return instrument;
-    }
-
-    public void setInstrument(byte instrument) {
-        this.instrument = instrument;
-    }
-
-    public byte getKey() {
-        return key;
-    }
-
-    public void setKey(byte key) {
-        this.key = key;
-    }
+	
+	private byte instrument;
+	private byte key;
 }
