--- src/main/java/com/xxmicloxx/NoteBlockAPI/PositionSongPlayer.java
@@ -1,41 +1,0 @@
-package com.xxmicloxx.NoteBlockAPI;
-
-import org.bukkit.Location;
-import org.bukkit.entity.Player;
-
-public class PositionSongPlayer extends SongPlayer {
-
-    private Location targetLocation;
-
-    public PositionSongPlayer(Song song) {
-        super(song);
-    }
-
-    public Location getTargetLocation() {
-        return targetLocation;
-    }
-
-    public void setTargetLocation(Location targetLocation) {
-        this.targetLocation = targetLocation;
-    }
-
-    @Override
-    public void playTick(Player p, int tick) {
-        if (!p.getWorld().getName().equals(targetLocation.getWorld().getName())) {
-            // not in same world
-            return;
-        }
-        byte playerVolume = NoteBlockPlayerMain.getPlayerVolume(p);
-
-        for (Layer l : song.getLayerHashMap().values()) {
-            Note note = l.getNote(tick);
-            if (note == null) {
-                continue;
-            }
-            p.playSound(targetLocation,
-                    Instrument.getInstrument(note.getInstrument()),
-                    (l.getVolume() * (int) volume * (int) playerVolume) / 1000000f,
-                    NotePitch.getPitch(note.getKey() - 33));
-        }
-    }
-}
