+++ src/main/java/com/xxmicloxx/NoteBlockAPI/Song.java
@@ -1,79 +1,45 @@
 package com.xxmicloxx.NoteBlockAPI;
 
+import com.google.common.collect.Maps;
+import lombok.Getter;
+
 import java.io.File;
 import java.util.HashMap;
 
+@Getter
 public class Song {
-
-    private HashMap<Integer, Layer> layerHashMap = new HashMap<Integer, Layer>();
-    private short songHeight;
-    private short length;
-    private String title;
-    private File path;
-    private String author;
-    private String description;
-    private float speed;
-    private float delay;
-
-    public Song(Song other) {
-        this.speed = other.getSpeed();
-        delay = 20 / speed;
-        this.layerHashMap = other.getLayerHashMap();
-        this.songHeight = other.getSongHeight();
-        this.length = other.getLength();
-        this.title = other.getTitle();
-        this.author = other.getAuthor();
-        this.description = other.getDescription();
-        this.path = other.getPath();
-    }
-
-    public Song(float speed, HashMap<Integer, Layer> layerHashMap,
-                short songHeight, final short length, String title, String author,
-                String description, File path) {
-        this.speed = speed;
-        delay = 20 / speed;
-        this.layerHashMap = layerHashMap;
-        this.songHeight = songHeight;
-        this.length = length;
-        this.title = title;
-        this.author = author;
-        this.description = description;
-        this.path = path;
-    }
-
-    public HashMap<Integer, Layer> getLayerHashMap() {
-        return layerHashMap;
-    }
-
-    public short getSongHeight() {
-        return songHeight;
-    }
-
-    public short getLength() {
-        return length;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public File getPath() {
-        return path;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public float getSpeed() {
-        return speed;
-    }
-
-    public float getDelay() {
-        return delay;
-    }
+	
+	private HashMap<Integer, Layer> layerHashMap = Maps.newHashMap();
+	
+	private short songHeight, length;
+	private String title;
+	private File path;
+	private String author, description;
+	private float speed, delay;
+	
+	public Song(Song otherSong) {
+		this.speed = otherSong.getSpeed();
+		delay = 20 / speed;
+		this.layerHashMap = otherSong.getLayerHashMap();
+		this.songHeight = otherSong.getSongHeight();
+		this.length = otherSong.getLength();
+		this.title = otherSong.getTitle();
+		this.author = otherSong.getAuthor();
+		this.description = otherSong.getDescription();
+		this.path = otherSong.getPath();
+	}
+	
+	public Song(float speed, HashMap<Integer, Layer> layerHashMap,
+	            short songHeight, final short length, String title, String author,
+	            String description, File path) {
+		this.speed = speed;
+		delay = 20 / speed;
+		this.layerHashMap = layerHashMap;
+		this.songHeight = songHeight;
+		this.length = length;
+		this.title = title;
+		this.author = author;
+		this.description = description;
+		this.path = path;
+	}
 }
