--- src/main/java/com/xxmicloxx/NoteBlockAPI/SongEndEvent.java
@@ -1,26 +1,0 @@
-package com.xxmicloxx.NoteBlockAPI;
-
-import org.bukkit.event.Event;
-import org.bukkit.event.HandlerList;
-
-public class SongEndEvent extends Event {
-
-    private static final HandlerList handlers = new HandlerList();
-    private SongPlayer song;
-
-    public SongEndEvent(SongPlayer song) {
-        this.song = song;
-    }
-
-    public static HandlerList getHandlerList() {
-        return handlers;
-    }
-
-    public SongPlayer getSongPlayer() {
-        return song;
-    }
-
-    public HandlerList getHandlers() {
-        return handlers;
-    }
-}
