+++ src/main/java/com/xxmicloxx/NoteBlockAPI/enums/FadeType.java
@@ -1,4 +1,4 @@
-package com.xxmicloxx.NoteBlockAPI;
+package com.xxmicloxx.NoteBlockAPI.enums;
 
 public enum FadeType {
     FADE_LINEAR
